login e cadastro de novo cliente

+++ auth/src/main/java/com/fluytcloud/auth/entities/Company.java
@@ -1,4 +1,4 @@
 package com.fluytcloud.auth.entities;
 
-public record Company(Long id, String name, String identifier) {
+public record Company(Integer id, String name, String identifier) {
 }
